Prevent unecessary re-render of user page.
Previously our posts for user selector was returning a new array every
time, and as useSelector gets called every time an action is dispatched,
any dispatched action would result in a re-render of the UserPage
component.

Created a memoized selector function using reselect which only returns a
new value if the arguments to it change. So when it's called again and
unrelated state has changed, such as notifications, the re-render does
not occur.