Stop unecessary re-renders of PostExcept.
Previously every PostExcerpt was being re-rendered when you click any
reaction button. This was because it caused a change in the posts array,
which caused PostsList to re-render, which caused all children to
re-render.

Fix is simple in this case, just React.memo PostExcerpt so it only
re-renders when its props actually change.