Change PostsList to use getEntityAdapter style.
Now PostsList renders posts using the ID list instead of the posts ID.
This means if a post changes in the store it doesn't cause a re-render
of the PostsList component, that only happens if a post is
added/removed/re-sorted.

Changed the PostExcerpt to use useSelector instead of be a plain
component. Because the parent PostsList component no longer re-renders
when a reaction is clicked, neither do PostExcerpts. And because they
call useSelector selecting only their own post, they only re-render when
their post changes.