Change add post to use RTK Query.
Note in AddPostForm we no longer need to manage setting requests state
ourselves.

Temporarily added in a refresh button which we can call to allow users
to refetch posts. This will allows us to see our updated post. Note this
demonstrates the use if isLoading vs. isFetching, where when we're
already loaded but fetching we know we have cached data, so we know in
this case we want to keep showing existing data with a loading
indicator. As opposed to when loading from fresh when we want a big
loading indicator as the main content.